refactor: receiver js
